+++ src/Traits/SoftDelete/SoftDeleteBehavior.php
@@ -18,16 +18,29 @@ class SoftDeleteBehavior extends RepositoryBehavior
      */
     private $newValue;
 
+    private ?bool $emptyValue;
+
     private bool $filterDeletedRows;
 
     /**
      * @param bool|DateTimeInterface $newValue use true if column is bool, null for DateTimeImmutable
+     * @param bool|null $emptyValue  override for "not deleted" value (undefined = use default behavior)
      */
-    public function __construct(string $column, $newValue = null, bool $filterDeletedRows = true)
-    {
+    public function __construct(
+        string $column,
+        $newValue = null,
+        bool $filterDeletedRows = true,
+        ?bool $emptyValue = null
+    ) {
         $this->column = $column;
         $this->newValue = $newValue ?? new DateTimeImmutable();
         $this->filterDeletedRows = $filterDeletedRows;
+
+        if (func_num_args() >= 4) {
+            $this->emptyValue = $emptyValue;
+        } else {
+            $this->emptyValue = $newValue === true ? false : null;
+        }
     }
 
     public function shouldFilterDeleted(): bool
@@ -46,6 +59,11 @@ class SoftDeleteBehavior extends RepositoryBehavior
     public function getNewValue()
     {
         return $this->newValue;
+    }
+
+    public function getEmptyValue(): ?bool
+    {
+        return $this->emptyValue;
     }
 
     public function withoutFilter(): self