+++ src/Traits/SoftDelete/SoftDeleteEventSubscriber.php
@@ -21,7 +21,7 @@ class SoftDeleteEventSubscriber extends EventSubscriber
         /** @var SoftDeleteBehavior $behavior */
         $behavior = $event->getRepository()->getBehaviors()->get(SoftDeleteBehavior::class);
         if ($behavior->shouldFilterDeleted()) {
-            $event->getQuery()->where($event->getRepository()->getTableName() . '.' . $behavior->getColumn(), null);
+            $event->getQuery()->where($event->getRepository()->getTableName() . '.' . $behavior->getColumn(), $behavior->getEmptyValue());
         }
         return $event->handle();
     }