+++ CHANGELOG.md
@@ -6,7 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Table alias which is used for class name
 
+### Fixed
+- Default value for SoftDeleteBehavior
 
 
 ## [0.5.8] - 2025-11-17